Add license to individual files Also split up algorithms borrowed from 254 into separate files
